requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.24.12
botocore==1.27.12
Django==4.0.5
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.1
Pillow==9.1.1
pycodestyle==2.8.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
